requirements.txt
streamlit
tensorflow
numpy
pandas
altair
pillow
jupyter
opencv-python